Fix confusing wording
Pun very much intended: I "grew up" with the Win32 API, where a WORD is
a 16-bit unsigned integer, and I named the functions that read uint16_t's
accordingly. This is not in line with the conventional meaning of the
term "word", so it's confusing. This commit fixes the problem.